Avoid using a vtkProp in interactor style.
vtkInteractorStyleGeoMap uses an actor to render the selection rectangle in
a nicer way (although presumably less performant given that the entire scene is
rendered while growing the area).  This approach does not fit directly the layer-
per-renderpass new paradigm since only actors in a layer would be rendered (this
actor never being added to a renderpass), hence a top layer (containing the
selection actor) would need to be added while doing rubber band selection.

Because of this, InteractorStyleGeoMap will default for now to its base class
approach which does not renders anything through the vtkRenderer but rather just
does it in the cpu and caches the current frame to restore it after selection
is done.