Fixed draw-polygon-selection for vtkPolydataFeatures.
Some features in databrowser (vtkPolydataOutlineFeature) maintain an internal
mapping of primitives different from the display id captured by vtkHardwareSelector.
Because of this, the picked features would not match the actual selection. To address
this, vtkPolydataFeature's API was expanded to provide a mapping between the picked
ids and the local ids.

Fixed help tags in example applications.  Added missing include. Clean-up.